chore : move annotation classes to annotation package
