+++ jodag-spring/src/main/java/jodag/annotation/EnableJodag.java
@@ -1,4 +1,4 @@
-package jodag;
+package jodag.annotation;
 
 import org.springframework.context.annotation.ComponentScan;
